refactor: refactor LearnMoreButton with CommonCtaButton
