+++ src/components/ui/index.ts
@@ -23,8 +23,8 @@ export type { ReactMarkdownWrapperProps } from "./ReactMarkdownWrapper";
 export { SecureYourSpot } from "./SecureYourSpot";
 export type { SecureYourSpotProps } from "./SecureYourSpot";
 export { JoinDiscordButton } from "./JoinDiscordButton";
-export { LearnMoreButton } from "./LearnMoreButton";
-export type { LearnMoreButtonProps } from "./LearnMoreButton";
+export { CommonCtaButton } from "./CommonCtaButton";
+export type { CommonCtaButtonProps } from "./CommonCtaButton";
 export { MailchimpSignupForm } from "./MailchimpSignupForm";
 export { ViewJobsScrollButton } from "../career/ViewJobsScrollButton";
 export type { ViewJobsScrollButtonProps } from "../career/ViewJobsScrollButton";
